Thread_id writeback for FLUs (mult)
Pipelined the thread_id that comes in `fu_data_t` in both the multiplier
(simple sequential 1 cycle multiplication) and in the serdiv
(pipelining until it's finished). From there, propagated into the output of
the `mult.sv` module in order to assign it to the writeback port `flu_thread_id_o`
of the EX stage when `mult_valid` is asserted.